backend/pyproject.toml
[project]
name = "chat-backend"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "python-dotenv>=1.0.1",
    "litellm>=1.75.0",
    "python-multipart>=0.0.12",
    "supabase>=2.10.0",
    "pydantic>=2.10.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.4",
    "pytest>=8.3.4",
]